requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
bleach==6.1.0
python-magic==0.4.27
openpyxl==3.1.2
flasgger==0.9.7.1
APScheduler==3.10.4
boto3==1.34.0
